test(repositories): ignore MongoRepositoryTests and EFCoreRepositoryTests
